♻️ [refactor]#11 logging -> log 폴더명 변경
