+++ src/main/java/profect/group1/goormdotcom/common/log/LoggingUtil.java
@@ -1,4 +1,4 @@
-package profect.group1.goormdotcom.common.logging;
+package profect.group1.goormdotcom.common.log;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
